Emit more error info for > 300 responses
The error message that is emitted and used in the error callback
function of the _request method only prints the status number. Flowdock
API provides much more useful context which is currently not surfaced
and is very useful for debugging.

In this commit:

1. cb() error callback is passed 'body' and 'res' parameteres of the
failed request in addition to the 'Error' object
2. Transpiled JS is included in the git repository, to facilitate testing
without publishing to npm
3. The 'Error' object prints out more information, like the
request/response bodies and paths.

NOTE: Change 1. will be sufficient to provide error information to the
users of this module. Change 2. can be reverted if we publish the flowdock
module under resin-io-playground and change 3. can be reverted when we debug
the syncbot bug that causes the error to be swallowed.